requirements.txt
customtkinter>=5.2.0
yt-dlp>=2024.0.0
pillow>=10.0.0
plyer>=2.1.0
requests>=2.31.0